add checking file existence

+++ src/main/java/FileIO/FileIO.java
@@ -123,7 +123,9 @@ public class FileIO {
     private static void renameFile(final String fileAddress, final String fileName, final String newName, final String type) {
         String path = fileAddress;
         File oldFile = new File(path, fileName + type);
-        File newFile = new File(path, newName + type);
-        oldFile.renameTo(newFile);
+        if (oldFile.exists()) {
+            File newFile = new File(path, newName + type);
+            oldFile.renameTo(newFile);
+        }
     }
 }
